Implement Data API 2.0.
While this does remove all previously implemented
DataProcessors, it was necessary to do so while the
foundation of the new implementation is made. With
DataManipulators using Keys, it was necessary for some
automatic retrieval of methods to both get and set
data without delegating to ValueProcessors.

To summarize, DataManipulators still store their
associated values, however, they now register
the field Getters and Setters AND their Value Getters.
It is integral from all implemented DataManipulators to
work through this as no extra API/implementation calls
are made.

ValueProcessors are specifically handling a certain
value type for a specific Key. In short, for each Key
made available through the API, there must be a
ValueProcessor associated with it. This still goes with
DataManipulators and DataProcessors.

While handling mutable ValueContainers are always easy,
a caching system was created for the use of retrieving
cached ImmutableValue(s) and ImmutableDataManipulator(s)
without requiring pre-generation. In other words, the
ImmutableDataCache is lazily loaded through various
forms of reflection and Guava's cache objects.

Finally, registration for both Keys and aforementioned
processors has been simplified in the game registry.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>