Implement Text placeholder API
Appropriate mixins and implementations are updated to implement
TextRepresentable as necessary.

Additionally, the json and XML formats have been updated to allow
specifying placeholders.

- For XML, the tag is <placeholder key="{key}" />
- For JSON, the additional key is placeholderKey

The two formats take fairly different approaches to specifying fallback
values, due to the fact that the JSON format has to remain compatible
with vanilla clients.

For XML, the fallback text, if any is present, is specified as a child
of the placeholder tag. No default placeholder text is included.

For JSON, the fallbackAsExtra parameter is set to true and the fallback
text is the first element in extra. If no fallback text is specified,
the default placeholder is included in the text field (which is ignored
when deserializing a component)

Closes #108