Implement handling of ip-related contexts.
Closes SpongePowered/Sponge#239.

This is implemnted in the contexts
- remoteip: The ip of the client
- localhost: The hostname the client used to connect to this server
- localip: The ip that received the client's connection on the server
- localport: The port number the server received this connection on

The two contexts remoteip and localip will detect ips matching those
configured in the `ip-sets` section of the global configuration (this
is only available globally). This section is a mapping of ip set names
(used as the context value) to a list of ip sets.

An example of the configuration:
```hocon
ip-sets {
    hype=[
        "fc00::/8"
        ]
}
```