Fix PlayerInteractBlockEvent triggering twice. Closes SpongePowered/Sponge#260
As per Grum :

This is a horrible hack needed because the client sends 2 packets on
'right mouse click' aimed at a block.
We shouldn't need to get the second packet if the data is handled but we cannot
know what the client will do, so we might still get it.
If the time between packets is small enough, and the 'signature' similar, we
discard the second one. This sadly has to remain until Mojang makes their
packets saner.