Move initUuid to earlier in the processLoginStart method.
This commit affects the Bungee IP Forwarding plugin only.

Prior to this commit, initUuid was injected at the same point as the fireAuthEvent method. However, fireAuthEvent was fired first, and as a result, no Game Profile was available to the ClientConnectionEvent.Auth event. By moving the injection to earlier in the processLoginStart method (just *after* we have the loginGameProfile field populated), we ensure that the spoofed profile is available before the event is fired.