WIP on Inventory Impl.
We start things off with implementing Slot. Basically I am abusing Java 8
so I can have a common interface between Mojang's slot and the inevitable
proxy Slot we need for creatures and such that have held equipment but
don't function like an Inventory does for a Player.

Tons of work to go...

Signed-off-by: Chris Sanders <zidane@outlook.com>