Perform IP Forwarding tasks earlier in the handshake process.
In order to be able to use the latest iteration of IP Forwarding for Bungee and Forge clients and thus achieve zero configuration on the Bungee side without breaking backwards compatibility, we need to be able to read the LoginProfile properties much earlier than we do right now, so we can indicate to SpongeForge that the user is a Forge client with zero config.

This does not impact on the existing implementation.