Ignore C10PacketCreativeInventoryAction packets during client full syncs.
Fixes SpongePowered/SpongeForge#426

This is another bad hack required since the client sends a C10 packet
for every slot containing an itemstack after a C16 packet in the following
scenarios :

1. Opening creative inventory after initial server join.
2. Opening creative inventory again after making a change in previous
inventory open.

This is done in order to sync client inventory to server and would be fine
if the C10 packet included an Enum of some sort that defined what type of sync
was happening.