Do not force set server gamemode on existing worlds.
For existing worlds, we set gamemodes on them ourselves but the
codesnippet I've changed forced it upon them from the server's mode in
server.properties. This change isn't breaking, the server's gamemode is
applied when a GameType is NOT_SET.

Further work needs to be done on graduating the gamemode to world configs.

Signed-off-by: Chris Sanders <zidane@outlook.com>