Fix distinction between dimType.getId and provider.getDimensionId.
For those that find the terms odd, they are but one must remember that
a dimensionType has an id and an instance of a WorldProvider having an id.

In reality, all worlds have an integer id and that is what makes a world
different, on the disk, fundamentally different between another world of
with the same dimension type.

Vanilla doesn't have this concept only because Vanilla has three worlds
all with completely different dimension types.

Signed-off-by: Chris Sanders <zidane@outlook.com>