Only process INTERACT_AT packets when handling UseEntity packets.
Fixes #632

The client will send 2 packets when right-clicking on any entity that is
not an armor stand. This causes 2 events to be fired for the same
right-click action. To workaround this issue, all INTERACT packets will be
ignored in favor of INTERACT_AT packets which are identical yet also
include the hit location.

As for knowing which interact method to run, we will first process the
interactAt method as this packet always is received first. If interactAt
returns false then we run the interactWith method.

The above changes allows our 'InteractEntityEvent.Secondary' event to
handle all right-clicks instead of requiring 2 separate events for
special-case entities such as armor stands.