Specify Locale.ENGLISH when using String#toLowerCase internally
Previously, internal uses of String#toLowercase did not pass in a
Locale, resulting in the system default being used. Some languages, such
as Turkish, end up mapping certain upercase ASCII characters (such as
"I") to a character other than the lowercase ASCII form of the same
character ("Ä" in Turkish).

This caused strings to be incorrectly compared when looking up IDs for CatalogTypes,
since all of the other internal uses of ids use the English lowercase
form.

This commit passes in Locale.ENGLISH when using String#toLowerCase in
instances such as CatalogType lookup. It does *not* change the behavior
for completely user-provided strings, such as in command parsing.