Implement Chunk#getNeighbor and optimize World#isAreaLoaded
Implement a version of Chunk#getNeighbor that uses neighbor pointers
that are updated on chunk load/unload. We then use this implementation to
create an optimized version of World#isAreaLoaded. This removes a major
source of chunk map lookups as it is called every tick for every entity and
was doing 25 chunk map lookups for each call.