Correct worlds having a split map/structure storage.
It is commonly accepted by modders that the location of the map storage is
a global data store. Sponge accidentally broke this assumption by giving
each World its own map storage (as well as a few other things). This
brings back in-line with Vanilla/Forge by ensuring the map storage is the
global one.

Also, pay no mind to the level of hacks in this commit. The EnderDragon
demanded it (here be dragons, blah blah blah).

Signed-off-by: Chris Sanders <zidane@spongepowered.org>