Fix some redstone blocks firing block place events instead of modify.
Fixes #726

Previously, setBlockState would check if the original and new block were different
and if so would capture as a place. The issue with this is minecraft can still use
the same block class but register 2 different blocks such as with redstone. To avoid
this issue, setBlockState will instead compare both block classes. If both classes
are the same, the block will be captured as a 'ChangeBlockEvent.Modify' instead of a
place.