Refactor configuration.
* Add per-world config option 'auto-save-interval' which is the tick interval used to save all loaded
chunks in a world. Set to a value 0 or less to disable.

* Add logging config option 'world-auto-save' which is used to log an entry when a world auto-saves.
Note: This setting is disabled by default as it can be spammy.

* Add global config option 'auto-player-save-interval' which is the tick interval used when
saving global player data. Set to a value 0 or less to disable.
Note: By default, all world data will save every 900 ticks (45 seconds) which is the vanilla
default.