Add special cause for drops caused by inventory close.
Plugins should be aware that cancelling any item spawn containing a named
cause "InventoryClose" will NOT restore the item properly to the source.
For example, if a player opens their own inventory and clicks an item onto
the mouse cursor then presses escape, if this gets cancelled then the item
that was originally on cursor will be lost for good.