Don't allow PlayerInstance's to queue chunks for unload.
We now remove the ability for a PlayerInstance to queue chunks for unload to
prevent chunk thrashing where the same chunks repeatedly unload and load.
This is caused by a player moving in and out of the same chunks.

Instead, the Chunk GC will now be responsible for going through loaded chunks
and queuing any chunk where no player is within view distance or a spawn chunk
is force loaded. With this change, it is highly recommended not to disable
the chunk GC as it is now fully responsible for unloading chunks. If you
find that your server is consuming more resources, try lower the default
gc tick interval which is currently at 600 ticks (30 seconds).