Optimize chunk and block ticks with neighbor checks.
Before a chunk is ticked, the chunk is checked if still loaded and if not
gets removed from active chunk set to prevent an unnecessary chunk tick.

Before a block is ticked, the surrounding chunk neighbors are now required to
be loaded. This prevents unnecessary chunk loads as well as updateTick
logic from running.

* Check surrounding chunk neighbors during checkForLight instead of only
the current chunk.
* Only return EmptyChunk's during spawner and world gen.