Add world config option 'leaf-decay'.
Due to the nature of leaf decay, it can become heavy on performance depending
on the amount of leaves active in chunks. To avoid any losses in
performance due to decay, disable leaf decay by setting 'leaf-decay' to false in
the world config.

Note: Disabling this will go against what vanilla intended so only turn it
off if you prefer performance.