Add last known username cache that never expires.
For tracking purposes, we generally only care about the last known
username when constructing a User to pass to events. This is the main
reason why we don't just use vanilla's cache since it expires by default
every 30 days. Depending on the amount of user cache being expired, it had
the potential to cause a big performance hit during refreshes.

Note: The SpongeUsernameCache is used to support both SpongeForge and
SpongeVanilla. As it is a shadow of Forge's version, you can find the
original code here :

https://github.com/MinecraftForge/MinecraftForge/blob/1.8.9/src/main/java/net/minecraftforge/common/UsernameCache.java

The only change we make is injecting into vanilla's PlayerProfileCache so
anything being added to vanilla's cache is also added to the username
cache.