Add GameProfileQueryTask to handle uuid lookups for tracking.
As chunks are loaded, any tracked UUID that needs a corresponding username
will now be queued in the GameProfileQueryTask. This task runs async every
second and will lookup a maximum of 2 UUID's per second. This is to avoid
getting blocked by mojang for too many requests.

As this task runs async, any updates to user cache must be thread-safe.
The PlayerProfileCache LinkedList has been changed to a
ConcurrentLinkedQueue and our SpongeUsernameCache HashMap has been changed
to a ConcurrentHashMap. This allows us to make changes to cache while
remaining thread-safe.