Ignore the chunk unloaded flag during unloads.
The chunk unloaded flag was added post 1.8.9 to avoid unloading a chunk if
a call was made to getLoadedChunk while queued for unload. While this may
be fine for vanilla servers, it goes against how we handle chunk unloads
in general. To avoid mods and/or plugins from having issues, we simply
ignore this flag and continue with pre 1.9 logic.