Re-introduce empty chunk to handle abnormal requests for chunk loads.
Many mods, including vanilla, tend to leak chunks which end up causing
extra processing for our Chunk GC as it has to constantly queue the same
chunks for unload everytime. The biggest offending methods are
getBlockState and getChunkFromBlockCoords. Both of these methods will load
or generate a chunk that does not exist.

To avoid this, we now validate the phase and/or state to determine whether
or not a chunk request should be allowed to load.
Note: This feature can be disabled by setting denyChunkRequests to false
in config.

* Fix ChunkGC adding duplicate chunks to unload queue.
* If chunk unload delay is configured, schedule unloads for chunk leaks.
  This prevents chunk leaks from immediately unloading at end of tick.

Fixes SpongePowered/SpongeForge#968
Fixes SpongePowered/SpongeForge#980

Conflicts:
	src/main/java/org/spongepowered/common/interfaces/world/gen/IMixinChunkProviderServer.java
	src/main/java/org/spongepowered/common/mixin/core/world/gen/MixinChunkProviderServer.java