Remove chunk unload queue.
Since we already loop through all loaded chunks to determine if a chunk is
scheduled for unload, we can simply check the unloaded flag and avoid the
queue entirely.

With all the new changes, the flow is as follows

1. If ChunkGC can process, all loaded chunks are scanned for leaks. Any
leaks that are found get marked for unload.
Note: The tick interval can be controlled by modifying the
 'chunk-gc-tick-interval' which by default runs every 600 ticks(30 seconds)

2. At end of world tick, each loaded chunk is checked to determine if it
should unload or not. There are 3 possibilities for a chunk not to unload
at end of a world tick :
 a. The chunk is persistent.
 b. The scheduled unload time has not been reached.
 c. The max unload limit has been reached.
Note: The scheduled unload delay only affects player chunks (not leaks).
This unload delay can be controlled by modifying the 'chunk-unload-delay'
setting in config. The default is 15 seconds which will reset if the chunk
becomes active again.

* Adjust chunk-gc-tick-interval default to 600 (30 seconds) for new logic.
* Adjust chunk-unload-delay default to 15.