Change whitelist mixin to subclass of UserListWhitelist
Normally all calls to the whitelist are redirected to the
WhitelistService that can be implemented by plugins. However, when
converting old whitelist save files this fails because the server
isn't fully initialized yet.

With these changes the implementation used for converting the old
whitelist save files stays the way it was designed in Vanilla and
doesn't throw an error.

Fixes #1043