Add tileentity activation feature.
For heavy mod servers that have thousands of tileentities, having each TE
perform updates while a chunk is loaded can be a significant burden on
performance. This feature gives servers the ability to control the range
and tick rate of every tileentity. TileEntities can be configured on a per
mod basis and per TE. This also allows servers to increase view-distance
without causing extra TE resources to be used.

Note: This feature is disabled by default and must be enabled via config.

The following configuration options have been added for tileentities under
the category "tileentity-activation" :

1. Block range required to tick. (default 64 blocks)

This checks the block range between a tileentity and players. If within
specified block range, the tileentity will perform its update logic.
If not within range, the update will be skipped.
Note: If a chunk is kept loaded with a "chunk loader", TE's will always
tick and ignore this check.

2. Tick rate. (default 1)

This controls how often a tileentity should tick. For example, if you
wanted furnaces to only update every other tick, you could set this to 2
instead of 1.
Note: If a chunk is kept loaded with a "chunk loader", this feature will
still take effect.