Fix block change cancellations not cancelling entity spawns. Fixes SpongePowered/SpongeForge#1219.
The way this works is that while certain phases are processing block changes,
those changes can lead to additional entity spawns that otherwise are
considered "single event spawns", leading to a lack of information
about what entities are actually spawned, causing duplication for items
dropped. The overall issue after the fact however, is that the entity
spawns are finally eliminated if the original block change that caused
those entities to spawn is cancelled. Now, this will appropriately propogate
the entity spawns according to the proper block position that is being
"processed" at the time, leading to only the desired entities to be
spawned according to plugins.

This should potentially require other generalized phase states to
follow suit in the cases of propogated block changes due to other
block changes in unwinding.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>