Change LensProvider to accept Fabric allowing passing ContainerFabrics. New ContainerLens used for Fallback when no Lens can be found for a Container. Use LensProviders for ContainerChest and ContainerPlayer.
When getting a Lens for a Container we first try to get existing Lenses from Adapter.
If the Container is no Adapter or has no Lens set yet we check for a LensProvider.
If there is still no Lens a new one will be generated for non-zero sized inventories:

For each inventory viewed in the Container we try to obtain a Lens and then wrap all of it in a ContainerLens.

fixes #1082