Require expected phase to be passed to CauseTracker#completePhase
Sometimes, the call to completePhase may end up being skipped, usually
due to Sponge failing to properly handle an exception. This phase may
end up getting unintentially popped by an unrelated call to
completePhase, (e.g. a PacketPhase causing a terrain gen phase to be
entered). While this is usually caught when Sponge tries to switch to a
phase that's incompatible with the phase erroneously left on the stack,
the exception is separated from the root cause of the problem - making
debugging a challenge.

To solve this, CauseTracker#completePhase now requires the 'expected'
phase to be passed in. This allows the cause tracker to fail fast in the
case where a call to completePhase has somehow been skipped.