Fix incorrect command permission checks in some situations.
When a Sponge plugin command takes a command alias that either Vanilla Minecraft or a Forge mod tries to take later, any permission checks on for the vanilla/mod command are directed to the Sponge plugin's command instead. This commit fixes this by:

* Adding updateNamespacedAlias(String ownerId) method to the CommandBase, which is called on instantiation of MinecraftCommandWrapper (and, by inheritance, ForgeMinecraftCommandWrapper), prefixing getName() with the owning plugin's ID.
* Redirecting the getName() call in checkPermission to use this namespaced alias, if it exists.

This then ensures that the permission check goes to the correct alias.

Fixes #1243