Don't capture block drops if fake player is cause.
ExtraUtilities 2's quantum quarry creates a special dimension to mine
blocks with a fake player and captures the block drops as it mines. In
order to make sure the captures work properly for mods such as Ex2, our
spawnAsEntity mixin will no longer capture if a fake player is detected with
an item in hand.

* Fix keepSpawnLoaded, shouldGenerateOnSpawn, loadOnStartup not properly
saving values to world config.
* Fix isMod detection in MixinWorldInfo.
* Validate loadOnStartup with DimensionType and force load if mod requires
it.