Implement Health Scaling.
The way this is implemented is that the client (and other clients) will see the scaled health, whereas
the server logic still retains the actual health. This way, mods and plugins, will not be falsely told
about the actual health value of a player. Scaling is also implemented in a way that it can be natively
removed (and therefor ignored) depending if a plugin wishes scaling to be removed in the future.
Since it persists, it now acts as a true data concept with processing only at the packet level to avoid
synchronization issues with alternative methods of implementation.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>