Add TestPlugins project into common for in-development only testing allowing a unified plugin for implementation testing in development environments.
What this specifically does is add the TestPlugins project such that implementation/api additions
made to the project can be simultaneously tested with the included plugin. Uniquely set up such
that the sources and output classes of said test plugin will not interfere with production jars
or setups. Ideally this will allow various commands to be created to test various features of
the API and implementation in various circumstances. This does require implementation updates
in SpongeForge and SpongeVanilla to function.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>