Don't use world from entity when spawning. Fixes SpongePowered/SpongeForge#830
If a mod attempted to spawn an entity into a world before calling
setWorld on the entity, the wrong world would be used as we were
assuming the entity was already set for target world. To fix the issue,
the world is now passed to our spawnEntity method to guarantee the
correct world is being used.