Fix wrong chunk folder being retrieved during mod init.
In a previous commit, SaveHandler#getWorldDirectory was altered for mods
that call this method directly. However, if a mod is being initialized and
chunks are being generated, the chunk loader will make a call in order to
find out which folder to place each region. To workaround this conflict,
we will now call an isolated method to guarantee the proper folder is
returned.

* Add tracking support for IEntityOwnable.
* Ensure chunks are loaded during spawn gen.

Fixes SpongePowered/SpongeForge#1699
Fixes SpongePowered/SpongeForge#1710