Couple of fixes for local movement.
The server defines local movement as any movement not within 4 blocks of
previous movement. As such this commit adjusts the API movement methods
to be aware if the movement being changed should be considered a
Teleport. This is only a stop-gap solution as the better fix will incur
a break (which will happen in a future commit).

One benefit of this change is we are no longer doing a chunk lookup for
movement within 4 blocks in this method as well as
detaching/re-attaching passengers (as that only happens in
teleportation).

Signed-off-by: Chris Sanders <zidane@spongepowered.org>