Disable the recipes plugin test.
The Forge client has a bug where if the server has a recipe and the
client does not, it will not be able to connect (it displays "Fatally
missing blocks and items!"). The Vanilla client does not have this
issue. In the interest of testing our code without setbacks, disabling
this plugin for now.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>