Couple of inventory event fixes. Updates #1448.
Keeping in mind a ContainerChest, which has a top and bottom
inventory...

- If you perform a drag across multiple inventories and only invalidate
transactions in one of those inventories, the cursor would not have the
correct quantity returned for the transactions invalided. This commit
resolves this.

- Likewise, if you perform a double click to gather up all items across
multiple inventories, the cursor would include additional quantity for
transactions who were actually invalidated (therefore duping them). This
commit resolves this.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>