Only allow custom beacon potions when set through the API
When a newly created beacon has the 'check mark' button pressed
(with fuel in the fuel slot) without clicking an effect button,
potion id 0 will be sent to the server. This will normally be rejected
by the server, since 0 isn't a valid vanilla potion id.

However, Sponge previously disabled this check in order to allow plugins
to set custom potions through the API. If potion id 0 is registered to a
mod potion effect, the disabled check will cause this to be set as the
beacon's primary effect when performing the process described above.

This commit restores Vanilla's original potion-checking behavior, except
when:
* The potion is being set through Sponge's Data API
* The potion is being restored from NBT.

This ensures that in-game interactions with the Beacon GUI work as
expected, without restricting the ability of plugins to set whichever
potion effect they choose.

Closes SpongePowered/SpongeForge#1493