It's working!!!!!
No but seriously.... this is working for entity flags, which
represents Keys.IS_SNEAKING, Keys.IS_SPRINTING, Keys.INVISIBLE,
and Keys.GLOWING. The nice thing is that this should be the fastest
we can churn out events in this circumstance.

Some testing still has to flow through to get some potion effects
handled correctly, since setting glowing on entities does not
get tracked through commands properly because it is applying an
effect, instead of setting the flag itself.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>