Only check for a TE when creating a SpongeBlockSnapshot.
Whenever a SpongeBlockSnapshot was generated, World.getTileEntity()
would be called in order to store TE NBT data if a TE existed. However,
getTileEntity would also attempt to create a TileEntity if it did not
exist. This caused mods to receive the original TileEntity during block
breaks when calling getTileEntity. To fix this issue,
createSpongeBlockSnapshot will now check for the TE in the chunk directly in
order to avoid creating a TileEntity.

One such example of this was caused in GiacomosFishingNet mod. After a
player attempted to break a fishing net in water, the mod would run
through its removedByPlayer logic. In its handler, it would call
getTileEntity and sponge would return the original TE instead of null causing
the mod to reset the blockstate.

Fixes SpongePowered/SpongeForge#1989