Don't reschedule block ticks for areas not loaded.
Blocks that sit on edge chunks (chunks who have no neighbors) may
trigger a scheduled update in that area not loaded. We patch the
isAreaLoaded check to not do chunk lookups (and instead neighbor
pointers) but now we'll have scheduled updates continuning forever (or
at the least until the neighbors of that chunk load). This commit
removes this concept.

Possible drawbacks may be automated farms on edge chunks but I find it
an acceptable compromise that edge chunks shouldn't be able to schedule
outside those bounds. If this becomes a problem, we can issue a
configurable option.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>

(cherry picked from commit 22ea118)