Force re-creation of SpongeUser object on login
Previously, a SpongeUser object for a given UUID would remain in the
user cache, until it expired. This lead to a number of issues:

* Until either the cache entry expired or the server restarted, the GameProfile
name of the user for a given UUID would remain fixed. If a player were
to change their name and rejoin the server, their EntityPlayerMP object
and SpongeUser object would have two different names.

* SpongeUser objects can be created before a player joins the server.
This can lead to a player's EntityPlayerMP gameProfile field storing a
different GameProfile object from their SpongeUser object.

Since the SpongeUser GameProfile is the GameProfile that's actually exposed
through the API, plugins would run into issues when attempting to modify
a player's GameProfile. Since Minecraft code (such as
SPacketPlayerListItem) uses the EntityPlayerMP gameProfile object,
changes made to the player's GameProfile by plugins would be not be
consistenly applied.

To resolve these issues, this commit forces the re-creation of a
player's SpongeUser object when they join the server. This keeps the
SpongeUser's name up-to-date, and ensures that the same GameProfile
object is used in EntityPlayerMP and SpongeUser.

SpongeUser objects are already periodically re-created, since they are
stored in an expiring cache. Re-creating SpongeUser objects on login
should therefore not cause any problems that weren't already present (e.g.
plugins using old SpongeUser objects).