Don't tick if chunk is in progress of being scheduled for unload.
In vanilla, the chunk is always unloaded immedately when no players are in
range so this is never a problem. However, in Sponge the chunk unload is
delayed by a default of 15 seconds when no players are in range. This
improves overall performance by preventing the same chunks from being
unloaded/loaded in a short period of time due to player movement.

In order to solve this issue and match vanilla behavior, entities and
TE's will no longer tick if the chunk is in progress of being scheduled
for unload.

Fixes SpongePowered/SpongeVanilla#344