Fix InteractItemEvent cancel not allowing block to activate.
If a protection plugin cancelled a player's item interaction on a block,
the block activation logic would never run. This was caused by our event
being called in a pre-stage before the block logic was even run. The
event has been moved to processRightClickBlock method in
PlayerInteractionManager in order to properly support item/block
results.