Make default activation state true for TE activation.
Previously, if TE activation was enabled, any TE not added to config
would not activate/tick since the default activation state was false.
This has been corrected and all TE's will now default to an active
state. The state will only be set to false if a block range has been
set that is greater than 0.