Actually copy custom data in ItemStackSnapshot.createStack().
Background: With ItemStackSnapshots, custom data is stored in three places: once in a list of ImmutableDataManipulators, once in the mutable DataManipulator list of a private ItemStack object, and once in the NBT data of the private ItemStack. However, the ItemStackSnapshot uses an additional NBT tag that doesn't include custom data. When converting back to a standard ItemStack, the NBT of the item stack is replaced with the NBT of the snapshot, which doesn't include custom data. This leads to an inconsistent state where the data is part of the resulting ItemStack's manipulator list but not part of its NBT. #1952 uncovered this through failed equality tests.

This commit fixes the inconsistent state of these item stacks by manually copying custom data after copying NBT data, which is guaranteed to add the custom data to both the NBT tag and the manipulator list.

Fixes #1952.