Include all captures in capture supplier toString(), not just counts
Now that we have config settings to limit phase tracker error spam,
including the full contents of a capture supplier shouldn't make much of
a difference. This can be valuable for debugging, especially when
dealing with recursive capture processing issues (where the captured
blocks may provide a hint as to the cause of the issue).