Experimental fix for now but something that will work definitevely. Fixes SpongePowered/SpongeForge#2273.
This fix is more of a quick "we know what the circumstances are" fix. Essentially, vanilla will allow
for an immediate block tick update to take place as necessary (water blocks placed by populators need this
so they will "flow", otherwise they just sit there doing nothing). The problem is that liquids do insanely
stupid flooding logic to where the flag being set on the world will enable all nested block updates to
take place immediately, thereby entering a massive recursive stack.

After lots of testing with this specific issue, the original issue causing a stack overflow is easily
reproducible with the SpongeForge added GrassToSponge testmod. SpongeForge only started recently triggering
the issue due to the deeply nested call stack to perform roughly the same logic that vanilla would run into.

The option for now is to immediately disable immediateBlockUpdates if we are processing an immediate block
update, so that any nested calls to update blocks will not be processed but scheduled for the next tick.

It's intended to make this a little more configurable such that a value option of "how many nested counts"
do we need to have before we shut off immediate block updates, but that will come later.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>