Fix bed respawn behavior near the world border
We were missing a check from Vanilla, which meant that we would loop
forever trying to increase the player's y position.

This commit also modifies the Vanilla bed check to take the worldborder
into account when determining if a position is valid. In Vanilla, a
player could sometimes be spawned in at the maximum height, standing on
top of the world border. With this check in place, Vanilla picks a
better position to begin with.