Prevent UserDiscoverer caching a user when a user doesn't exist
Prior to this change, the first call to UserStorageService#get for a profile that did not have an User data correctly returned an empty optional. However, this then created a user object and put it into the cache, so that on the second call to UserStorageService#get for the same profile returned a User. This commit moves the user creation so that this doesn't happen, while tracking UUIDs that did not have a file.

Now, UserStoarageService#get will produce the same result over and over unless a User is specifically created.

It does mean that you cannot manually add a user file while the server is running if the user has already been requested, but this was already the case anyway.

A test plugin is included to verify this behaviour.