Prevent plugins from accessing custom data on empty ItemStacks
Internally, Vanilla may temporarily set an ItemStack's size to 0, and
later set it back to a positive number. Because of this, we need to keep
manipulators stored in an ItemStack even when it's empty. However,
plugins should never be able to see manipulators on an empty ItemStack,
since all empty ItemStacks are 'the same' from an API perspective.

This commit adds isEmpty checks in any API-accessible methods that
access custom data. This prevents issues where custom data would appear
to linger in a particular inventory slot, even when the item in the slot
had been dropped.