Slightly change the flow for interaction events.
In the past, cancelling the InteractItemEvent would still allow the InteractBlockEvent to continue but would set the item use attribute of the InteractBlock to false. Except in reality this doesn't make sense logically if you consider:

Hold item (empty hand is an item) -> right click door -> cancel item interact -> door still opens(?).

Cause and effect says that if the held item's interact was cancelled, the door would have never opened. In the interest of being flexible, we will still fire the InteractBlock event but we'll just cancel it at the start. We also still honor if the InteractBlock states to cancel the item usage.

Signed-off-by: Chris Sanders <zidane@spongepowered.org>