Properly fix SpongePowered/SpongeForge#1946. Thanks to all who participated. The bug was uncovered partially with the previous commit targeting the NPE, but highlighted that the biomesForGeneration array was the null object, not the biome element within the array.
What uncovered the issue entirely was that the biomes appear to be null, only in the context
of running a server, the null does not show up if the world generation occurs in single player,
leading to the issue being unreproducible for developers. Attempting to ensure the array is
not null any time prior leads to the recreation of what Mojang originally wrote where the biomes
for generation are re-created for the particular area to avoid a null. I anticipate this will
be totally eliminated in 1.13.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>