Partially revert #1958 for world limit fix
A common source of confusion is the decimal co-ordinates vs block
co-ordinates. It is stated that the minecraft X/Z co-ordinates have
a range of +/-30,000,000, which is true, however, the block coordinates
range from -30,000,000 to +29,999,999.

This is because the edges of blocks line up with the unit co-ordinates,
such that a block spans from n to n+1 (where n is an integer). Thus, a
block that we class as being at (29,999,999, 0, 29,999,999) actually
spans to (30,000,000, 0, 30,000,000).

The block limit should be within the 30M radius around 0, meaning that
the last possible block that can be interacted with is at 30M-1, as the
block with block co-ordinate  X=30M actually strays past the Minecraft
border to decimal co-ordinate 30M+1 - which is out of bounds.

This commit reverts the max block limit change.