Lord of the Blocks: The Fellowship of the Tracker Introduce BlockPosMultiMap. The context of its usage is to replace the previous existing CaptureObject that maintained a List of BlockSnapshots such that this new Multimap will retain a set of BlockPoses to maintain a List of intermediary BlockSnapshot changes for each BlockPos.
The reasoning for this is that certain case of phases will
have multiple changes per block position and require
processing of said positions. The issue is that while these
multiple changes need to be tracked, the capturing would
be excessive for cases where multiple changes per position
do NOT take place. This specific ICaptureSupplier is designed
to retain the same footprint of the original block capture
object, but with the advantage of dynamically changing over
to a Multimap based implementation when necessary.

The costs for processing and viewing are yet to be determined,
however the implementation is still needing to be updated for
the changes as processing needs to reflect the order in which
block changes take place, while still processing the
ChangeBlockEvent based on the original BlockSnapshot that existed
compared to the BlockSnapshot that is taking it's place.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>