Ensure that the cause and context for all UseItemStackEvent events is consistent
Some UseItemStackEvent sub-events (e.g. UseItemStackEvent.Start) are
fired from packet handlers, which causes the player to already be set in
EventContextKeys.OWNER and EventContextKeys.NOTIFIER.

However, other sub-event, such as UseItemStackEvent.Tick, are not fired
in response to packets, causing them to have an empty context.

To ensure that plugins do not need to care about this distinction, this
commit explicitly adds the player to the cause and context for all
UseItemStackEvent sub-events

Fixes #2166