Lord of the Blocks: The Five Classes
Documenting the uses of PhaseContext#getCapturedBlocks vs using the
MultiBlockCaptureSupplier. Started refactoring the processing of captured
blocks such that TrackingUtil no longer uses a passed object that would
otherwise be retrieved from the PhaseContext being passed in to begin
with.

Added many todo's for handling individual block snapshot removal, and the
consideration of changes to be made for Transaction as a class.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>