Lord of the Blocks: The Road to Eventgard
Actually implemented most of MultiBlockCapture being processed, but
finding that it may be necessary to capture notification events as well as
tile entity creation and removal during block events to properly play them
back for proper cancellation, avoiding client issues. This sort of cost
should be minimal, once optimized and verified it works.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>