Fix scaled health stripping attribute modifiers.
When health scaling is set, the attribute modifiers that
mods add need to be shown as they can emphasize the health
is being added onto. This is tricky in situations where
the player health is in fact scaled, but the attribute modifier
is not. To allow proper representation of the modifiers,
the scale needs to be modified as well, to properly display
on clients when the health is being added on.

Likewise, by providing the attribute modifiers to the client,
this allows mods who keep track of those modifiers to properly
track what the new maximum health is for themselves, regardless
whether sponge is scaling the base health or not.

This change fixes SpongePowered/SpongeForge#2544.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>